Fixes related to the latest pr review

+++ Tensile/Source/lib/include/Tensile/PropertyMatching.hpp
@@ -495,7 +495,7 @@ namespace Tensile
                         model_K = iter->key[2];
                     }
                     bestDistance = theMatch->computeTAMScore(
-                        object, hardware, model_M, model_N, model_K, model_NumBatches);
+                        object, hardware, (double)model_M, (double)model_N, (double)model_K, (double)model_NumBatches);
                 }
 
                 iter++;
@@ -521,7 +521,7 @@ namespace Tensile
                             model_K = iter->key[2];
                         }
                         double nextDistance = theMatch->computeTAMScore(
-                            object, hardware, model_M, model_N, model_K, model_NumBatches);
+                            object, hardware, (double)model_M, (double)model_N, (double)model_K, (double)model_NumBatches);
 
                         if(nextDistance < bestDistance)
                         {
